Regenerate all contentpacks according to the new spec
Regenerate all contentpack ZIP files according to the new reproducible
build specification. This causes a mass regeneration, but it's
one-time only.

Moving forward, some version numbers might go back in time
unexpectedly if newest files are removed. Also, the version number
only captures the mtime of the latest file, so if old files are
removed the version number will not change. I don't know yet if this
is a big deal. I will wait and see.